+++ lib/ReactViews/Map/MenuBar.jsx
@@ -125,7 +125,8 @@ const MenuBar = createReactClass({
               terria={this.props.terria}
               viewState={this.props.viewState}
             />
-            {satelliteGuidancePrompted &&
+            {this.props.terria.configParameters.showFeaturePrompts &&
+              satelliteGuidancePrompted &&
               !mapGuidesLocationPrompted &&
               !this.props.viewState.showSatelliteGuidance && (
                 <Prompt